add non-working exhibition tests
